LPD-20574 Move Solutions Component to folder
