+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/components/DashboardTable/DashboardTable.tsx
@@ -37,21 +37,21 @@ export type TableHeaders = {
 interface DashboardTableProps<T> {
 	children?: (item: T) => ReactNode;
 	emptyStateMessage: {
-		description1: string;
-		description2: string;
+		description1?: string;
+		description2?: string;
 		title: string;
 	};
 	icon: string;
-	items: T[];
-	tableHeaders: TableHeaders;
+	items?: T[];
+	tableHeaders?: TableHeaders;
 }
 
 export function DashboardTable<T>({
 	children = () => null,
 	emptyStateMessage,
 	icon,
-	items,
-	tableHeaders,
+	items = [],
+	tableHeaders = [],
 }: DashboardTableProps<T>) {
 	if (items.length) {
 		return (
@@ -85,8 +85,8 @@ export function DashboardTable<T>({
 
 	return (
 		<DashboardEmptyTable
-			description1={description1}
-			description2={description2}
+			description1={description1 ?? ""}
+			description2={description2 ?? ""}
 			icon={icon}
 			title={title}
 		/>
